Backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
google-generativeai
pandas
psycopg2-binary
PyJWT
